refactor: hitOrStand 함수 분리
