+++ src/main/kotlin/blackjack/BlackjackApplication.kt
@@ -17,18 +17,7 @@ fun main() {
 
     players.forEach { player ->
         runCatching {
-            while (true) {
-                if (!player.canHit()) {
-                    println("${player.name}의 카드가 21 이상입니다. 카드를 더 받을 수 없습니다. \n")
-                    break
-                }
-                val hit = InputView.hitOrStand(player.name)
-                if (hit) {
-                    playGame.hit(player)
-                }
-                OutputView.cardOfPlayer(player)
-                if (!hit) break
-            }
+            playGame.hitOrStand(player)
         }.onFailure { println(it.message) }
     }
